### 2023-02-26 #5189 @editorialbot
[PRE REVIEW]: spiketools: a Python package for analyzing single-unit neural activity
**Submitting author:** @TomDonoghue (<a href="http://orcid.org/0000-0001-5911-0472">Thomas Donoghue</a>)
**Repository:** https://github.com/spiketools/spiketools
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.0
