### 2023-02-28 #5189 @AJQuinn
:wave:  Hello - @rly, @neuromusic, @djsaunde, @paxtonfitzpatrick, would any of you be willing to review this submission for JOSS? The code repository and paper details are further up in this thread.

We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html - please check them out if you haven't already.

Cheers!
