### 2023-03-17 #5268 @editorialbot
[REVIEW]: spiketools: a Python package for analyzing single-unit neural activity
**Submitting author:** @TomDonoghue (<a href="http://orcid.org/0000-0001-5911-0472">Thomas Donoghue</a>)
**Repository:** https://github.com/spiketools/spiketools
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.0
