### 2023-04-02 spiketools/spiketools#159 @rly
Suggested improvements / typo fixes in tutorials
Hi, I am reviewing `spiketools` as part of https://github.com/openjournals/joss-reviews/issues/5268. Great documentation and tutorials. I appreciate the effort put into both. In going through the tutorials, I found a few points of potential improvement and typos to address.

### https://spiketools.github.io/spiketools/auto_tutorials/plot_measures.html#sphx-glr-auto-tutorials-plot-measures-py:
1. The text "Again, we can see the converted data has the same raster plot as the original data." is not exactly accurate -- the converted data has the same raster plot as the original data, minus an offset.

### https://spiketools.github.io/spiketools/auto_tutorials/plot_spatial.html#sphx-glr-auto-tutorials-plot-spatial-py:
2. In the second cell
```python
bin_widths = np.array([1, 1, 1, 2, 1.5, 1, 1, 1, 2, 1, 1, 1, 1, 1, 0.5])
```
is defined but not used and is therefore confusing. I suggest removing it.

3. The `bins` variable defined in that cell is also not used until considerably later and might be confusing to define this early. I suggest moving the definition to right before it is used.

4. This code:
```python
# Compute speed at each point
bin_widths = np.array([1, 1, 1, 2, 1.5, 1, 1, 1, 2, 1, 1, 1, 1, 1, 0.5])
```
is confusing where it comes from. I suggest replacing it with `np.diff(timestamps)` which is equivalent and more clear.

5. The comment `# Compute the durations of the timestamps` does not make sense to me. "the durations of the samples` might be better.

6. I think in the comment `# Define speed threshold, used to position values for speed less than the threshold`, "position" should be changed to "remove" or "ignore" or "remove position".

### https://spiketools.github.io/spiketools/auto_tutorials/plot_stats.html#sphx-glr-auto-tutorials-plot-stats-py
7. The shuffle poisson raster looks like one spike train permuted. Indeed, the source code says `isis = permute_vector(compute_isis(poisson_spikes), n_permutations=n_shuffles)` and `permute_vector` says that there is no randomness - the permutation is just moving an ISI and a spike from the end to the beginning. If your num shuffles is small, then your random permutations will be very similar to each other while being quite different from the original data. Therefore, this would  not serve as a good baseline distribution for stats. I recommend one of the following: add a warning about this behavior in the function documentation, introduce randomness such that `poisson_generator` is called once for each shuffle instead of just once, introduce randomness in `permute_vector`.

8. In the code `shuffled_spikes = shuffle_spikes(spikes, 'CIRCULAR', shuffle_min=200, n_shuffles=10)`, I imagine that the 'CIRCULAR' arg is not case sensitive, but having it in all caps perhaps suggests that it might be? Not important, but just looks odd.

9. In the cell with `df_pre_post = create_dataframe(data)`, I think it would be useful to print `df_pre_post` in the same cell to give readers a better understanding of the table.

10. `Now that we have our data organized into a dataframe, we can run an ANOVA using.` - remove "using" at the end.

11. Next line `analyze whether there is an of the event on firing rates` - "an of" -> "an effect"

12. `This will give us a the surrogate distribution` - "a the" -> "a"

13. `To do so, we will simulating spike trains across trials (8 seconds) across spatial bins.` - "simulating" -> "simulate"

14. Like in point 9 above, in the cell with `df = create_dataframe_bins(bin_firing_all, dropna=True)`, I think it would be useful to print `df` in the same cell to give readers a better understanding of the table.
