### 2023-04-02 spiketools/spiketools#160 @rly
Documentation search bar doesn't work
Hi, I am reviewing spiketools as part of https://github.com/openjournals/joss-reviews/issues/5268.

https://spiketools.github.io/spiketools/search.html?q=test&check_keywords=yes&area=default
just says "Searching..."

The console reports:
```
searchtools.js:243 Uncaught ReferenceError: Stemmer is not defined
    at Object.query (searchtools.js:243:21)
    at Object.performSearch (searchtools.js:234:35)
    at HTMLDocument.init (searchtools.js:175:23)
```
