### 2023-04-22 4c19a92
add check_param_type function
