+++ spiketools/tests/utils/test_checks.py
@@ -53,6 +53,15 @@ def test_check_param_lengths():
     with raises(ValueError):
         check_param_lengths([a1, a2], ['a1', 'a2'], expected_length=2)
 
+def test_check_param_type():
+
+    param = [1, 2]
+
+    check_param_type(param, 'test', [list, np.ndarray])
+
+    with raises(ValueError):
+        check_param_type(tuple(param), 'test', [list, np.ndarray])
+
 def test_check_array_orientation():
 
     arr1 = np.array([1, 2, 3])
