### 2023-04-22 22c0f34
use check_param_type
