### 2023-04-22 b54195b
update plot_rate_by_time to support single 2d array
