### 2023-04-22 f48ed35
plot_rate_by_time - fix colors mgmt for str input
