### 2023-04-26 spiketools/spiketools#160 @TomDonoghue
Good catch! This appears to have been an issue with one of our docs dependencies - updating things seems to have resolved this issue!

Site has been re-deployed, and search now seems to work!
