### 2023-04-27 799094c
fix tests for plot_positions
