### 2023-04-27 43b2f07
add add_gridlines to style funcs
