+++ spiketools/tests/plts/test_style.py
@@ -1,5 +1,7 @@
 """Tests for spiketools.plts.style."""
 
+import numpy as np
+
 from spiketools.plts.settings import SET_KWARGS
 
 from spiketools.plts.style import *
@@ -48,3 +50,19 @@ def test_invert_axes():
     invert_axes('both', ax2)
     assert ax2.get_xlim()[0] > ax2.get_xlim()[1]
     assert ax2.get_ylim()[0] > ax2.get_ylim()[1]
+
+def test_add_gridlines():
+
+    _, ax1 = plt.subplots()
+    ax1.plot([0, 2], [0, 2])
+
+    bins = [0.5, 1.5]
+    add_gridlines(bins, None, ax1)
+    assert np.array_equal(ax1.get_xticks(), bins)
+    assert np.array_equal(ax1.get_yticks(), [])
+    add_gridlines(None, bins, ax1)
+    assert np.array_equal(ax1.get_xticks(), [])
+    assert np.array_equal(ax1.get_yticks(), bins)
+    add_gridlines(bins, bins, ax1)
+    assert np.array_equal(ax1.get_xticks(), bins)
+    assert np.array_equal(ax1.get_yticks(), bins)
