### 2023-04-27 1899700
use add_gridlines & remove extra ifs
