### 2023-04-27 8f0b853
update plot_position_by_time to pass through to plot_positions
