### 2023-04-27 b05bfbc
add support for plotting events in plot_position_by_tome
