### 2023-04-28 90b22ff
add listify func
