### 2023-04-28 ddc5524
use new listify function
