+++ spiketools/spatial/checks.py
@@ -1,5 +1,7 @@
 """Checker functions for spatial related functionality."""
 
+from spiketools.utils.base import listify
+
 ###################################################################################################
 ###################################################################################################
 
@@ -43,12 +45,10 @@ def check_spatial_bins(bins, position=None):
         Raised if there are any issues with the given bin definition.
     """
 
-    if isinstance(bins, int):
-        bins = [bins]
+    bins = listify(bins)
 
-    if isinstance(bins, list):
-        for binval in bins:
-            assert isinstance(binval, int), 'Bin definition values should be integers.'
+    for binval in bins:
+        assert isinstance(binval, int), 'Bin definition values should be integers.'
 
     assert len(bins) <= 2, 'Bin definition has too many values (>2).'
 
