+++ spiketools/plts/spatial.py
@@ -5,7 +5,7 @@ from copy import deepcopy
 import numpy as np
 import matplotlib.pyplot as plt
 
-from spiketools.utils.base import combine_dicts
+from spiketools.utils.base import listify, combine_dicts
 from spiketools.utils.checks import check_array_lst_orientation
 from spiketools.utils.data import make_row_orientation, smooth_data, compute_range
 from spiketools.modutils.functions import get_function_parameters
@@ -52,10 +52,8 @@ def plot_positions(position, spike_positions=None, landmarks=None, x_bins=None,
 
     ax = check_ax(ax, figsize=plt_kwargs.pop('figsize', None))
 
-    position = [position] if isinstance(position, np.ndarray) else position
-    orientation = check_array_lst_orientation(position)
-
-    for cur_position in position:
+    orientation = check_array_lst_orientation(listify(position))
+    for cur_position in listify(position):
         ax.plot(*make_row_orientation(cur_position, orientation),
                 color=plt_kwargs.pop('color', DEFAULT_COLORS[0]),
                 alpha=plt_kwargs.pop('alpha', 0.35),
