### 2023-04-28 b076ea1
add plot_position_1d
