### 2023-05-03 5c10fce
fix up check_param_type and usage in spatial checks
