### 2023-05-03 b594c97
update cumul dist to match length of input
