### 2023-05-03 61d0bc0
update docstring examples to be easier to check
