### 2023-05-04 daf6420
add align_output option to relevant position funcs
