### 2023-05-04 37cd3fb
updates & orgs of spatial checks
