+++ spiketools/tests/spatial/test_checks.py
@@ -20,7 +20,7 @@ def test_check_position():
     with raises(AssertionError):
         check_position(position3)
 
-def test_check_spatial_bins():
+def test_check_bin_definition():
 
     bins_int = 12
     bins_1d = [12]
@@ -28,16 +28,23 @@ def test_check_spatial_bins():
 
     position = np.array([[1, 2, 3], [4, 5, 6]])
 
-    out1d1 = check_spatial_bins(bins_int)
-    out1d2 = check_spatial_bins(bins_1d)
+    out1d1 = check_bin_definition(bins_int)
+    out1d2 = check_bin_definition(bins_1d)
     assert out1d1 == out1d2 == bins_1d
 
-    bins = check_spatial_bins(bins_2d, position)
+    bins = check_bin_definition(bins_2d, position)
     assert bins == bins_2d
 
     with raises(AssertionError):
-        check_spatial_bins([1.2, 2.3])
+        check_bin_definition([1.2, 2.3])
     with raises(AssertionError):
-        check_spatial_bins([1, 2, 3])
+        check_bin_definition([1, 2, 3])
     with raises(AssertionError):
-        check_spatial_bins(bins_1d, position)
+        check_bin_definition(bins_1d, position)
+
+def test_check_bin_widths():
+
+    check_bin_widths(np.array([1, 1, 1]))
+
+    with raises(AssertionError):
+        check_bin_widths(np.array([1, 1, 2]))
