### 2023-05-04 3f32993
update compute_speed to take in timestamps
