### 2023-05-04 019208b
move compute_sample_durations
