+++ spiketools/spatial/occupancy.py
@@ -8,8 +8,9 @@ import pandas as pd
 from spiketools.utils.data import assign_data_to_bins
 from spiketools.utils.checks import check_array_orientation
 from spiketools.utils.extract import get_values_by_time_range
+from spiketools.utils.timestamps import compute_sample_durations
 from spiketools.spatial.checks import check_position, check_bin_definition
-from spiketools.spatial.utils import get_position_xy, compute_sample_durations
+from spiketools.spatial.utils import get_position_xy
 
 ###################################################################################################
 ###################################################################################################
