### 2023-05-04 41d5d66
fix check_param_type for managing input types format
