### 2023-05-04 cadbdd4
add extract func testing for valid timepoint inputs
