### 2023-05-06 38e06bc
add plot_task_events to doc api
