### 2023-05-06 e6a500b
small fixes to plot_task_structure
