### 2023-05-11 spiketools/spiketools#171 @djsaunde
Examples link in README is broken
https://spiketools.github.io/spiketools/auto_examples/index.html gives a 404 error.
