### 2023-05-11 #5268 @djsaunde
[This issue](https://github.com/spiketools/spiketools/issues/171) is currently blocking " Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)."
