### 2023-05-11 c2a6fb5
Update README.rst
drop dead examples link
