### 2023-05-11 spiketools/spiketools#171 @TomDonoghue
Thanks for noticing / reporting this - broken link removed in #172
