### 2023-05-11 20df768
minor tutorial fixes
