### 2023-05-11 spiketools/spiketools#159 @TomDonoghue
Thanks a ton @rly for the notes here, these are super useful! All the tutorial specific notes and suggested were addressed in #164. For topic 7, about shuffle poisson, the issue here is a bit deeper than a tutorial element, and so we have moved that note to #163 to discuss / address / fix there.
