### 2023-05-11 #5268 @TomDonoghue
Hey all - first thing, thank you to everyone's work on this review, we greatly appreciate it!

@rly : thanks for your review, and suggestions you made - we have made those updates through the issues!

@djsaunde : thanks also for the review and for your comments. For the specifics of the spiketools issue you opened (https://github.com/spiketools/spiketools/issues/171) this has been addressed in a literal way by removing the broken link (there some notes [here](https://github.com/spiketools/spiketools/pull/172) contextualizing that that broken link was due to adapting materials from another project. For this project, our main set of examples are in the [tutorials](https://spiketools.github.io/spiketools/auto_tutorials/index.html), which covers examples of key sub-modules (as well as having doctest examples within a large number of function docstrings). For the purpose of the review, do you find the available materials sufficient for confirming the "example usage" topic, or do you mean that you think we need to add an additional examples page?

@AJQuinn - thank you for handling this submission!
I'm not sure if @neuromusic is still expected respond here, and if not, do we need to find another reviewer, or is two sufficient?
