### 2023-05-20 spiketools/spiketools#173 @neuromusic
suggestion: give users more guidance on installing optional dependencies
Its great that the Install page notes that `statsmodels` is an optional dependency, however, you can make this even easier for users, with one (or both) of the following:

1. Under the ["installation" section](https://spiketools.github.io/spiketools/#installation), add a second "optional" instruction to use `pip install statsmodels`
2. Alternatively, add `statsmodels` as an extra dependency for setuptools so that `pip install spiketools[all]` or the like will install all dependencies

(related to https://github.com/openjournals/joss-reviews/issues/5268)
