### 2023-05-20 spiketools/spiketools#174 @neuromusic
idea: add Binder support to help users kick the tires
Nice job using sphinx-gallery to give users downloadable, executable examples!

If you want to go a step farther, I'd recommend exploring configuring sphinx gallery to generate [Binder](https://mybinder.org/) links as well. This will let prospective users run your tutorials with one click (rather than needing to configure a local environment and download the examples).

There's more info on how do do this in the Sphinx-Galley docs here: https://sphinx-gallery.github.io/stable/configuration.html#generate-binder-links-for-gallery-notebooks-experimental

(related to https://github.com/openjournals/joss-reviews/issues/5268)
