### 2023-05-20 3d42a68
adds binder config to conf.py
first pass at getting a binder config added to the sphinx config
