+++ doc/conf.py
@@ -129,4 +129,16 @@ sphinx_gallery_conf = {
     'doc_module': ('spiketools',),
     'reference_url': {'spiketools': None},
     'remove_config_comments': True,
+    'binder': {
+        # Required keys
+        'org': 'spiketools',
+        'repo': 'spiketools',
+        'branch': 'v{}'.format(__version__), # Can be any branch, tag, or commit hash. Use a branch that hosts your docs.
+        'binderhub_url': 'https://mybinder.org', # Any URL of a binderhub deployment. Must be full URL (e.g. https://mybinder.org).
+        'dependencies': ['../requirements.txt', '../optional-requirements.txt'],
+        # Optional keys
+        # 'filepath_prefix': '<prefix>' # A prefix to prepend to any filepaths in Binder links.
+        # 'notebooks_dir': '<notebooks-directory-name>' # Jupyter notebooks for Binder will be copied to this directory (relative to built documentation root).
+        # 'use_jupyter_lab': <bool> # Whether Binder links should start Jupyter Lab instead of the Jupyter Notebook interface.
+     }
 }
