+++ binder/requirements-binder.txt
@@ -1,1 +1,0 @@
-spiketools
