+++ binder/requirements.txt
@@ -1,3 +1,6 @@
--r ../requirements.txt
--r ../optional-requirements.txt
+numpy
+pandas
+scipy
+matplotlib
+statsmodels
 spiketools
