### 2023-05-24 456a64c
add extract functions for ind(s) by value(s)
