+++ doc/api.rst
@@ -488,6 +488,8 @@ Utilities for extracting data segments of interest.
    create_mask
    get_range
    get_value_range
+   get_ind_by_value
+   get_inds_by_values
    get_ind_by_time
    get_inds_by_times
    get_value_by_time
