### 2023-05-24 #5268 @TomDonoghue
Thanks @neuromusic for the review and very useful comments! I have now addressed & merged in the suggested updates [https://github.com/spiketools/spiketools/issues/173 & https://github.com/spiketools/spiketools/issues/174].

@AJQuinn - I think everything is now addressed from the review side of things - is there anything else you need us to do on our side of things?
