### 2023-05-24 #5268 @AJQuinn
Fantastic, thanks all!

@neuromusic, @rly and @djsaunde - can I get a final confirmation that  you're happy and can sign off on `spiketools`?.
