### 2023-05-24 #5268 @rly
I'm happy with `spiketools` and give it the "go" signal.
