### 2023-05-25 d99dd8d
refactor position functions to handle pos array
