+++ spiketools/tests/spatial/test_position.py
@@ -9,53 +9,92 @@ from spiketools.spatial.position import *
 
 def test_compute_distance():
 
-    out0 = compute_distance(1, 1, 1, 1)
-    assert isinstance(out0, float)
-    assert np.isclose(out0, 0.0)
+    # 1d
+    out1d0 = compute_distance([1], [1])
+    assert isinstance(out1d0, float)
+    assert np.isclose(out1d0, 0.0)
 
-    out1 = compute_distance(1, 1, 2, 1)
-    assert isinstance(out1, float)
-    assert np.isclose(out1, 1.0)
+    out1d1 = compute_distance([1], [2])
+    assert isinstance(out1d1, float)
+    assert np.isclose(out1d1, 1.0)
 
-    out2 = compute_distance(1, 1, 2, 2)
-    assert np.isclose(out2, np.sqrt(2))
+    # 2d
+    out2d0 = compute_distance([1, 1], [1, 1])
+    assert isinstance(out2d0, float)
+    assert np.isclose(out2d0, 0.0)
 
-    out3 = compute_distance(1, 1, -1, 1)
-    assert np.isclose(out3, 2)
+    out2d1 = compute_distance([1, 1], [2, 1])
+    assert isinstance(out2d1, float)
+    assert np.isclose(out2d1, 1.0)
+
+    out2d2 = compute_distance([1, 1], [2, 2])
+    assert np.isclose(out2d2, np.sqrt(2))
+
+    out2d3 = compute_distance([1, 1], [-1, 1])
+    assert np.isclose(out2d3, 2)
 
 def test_compute_distances():
 
-    xs = np.array([0, 0, 1, 1, 2])
-    ys = np.array([0, 0, 0, 1, 2])
+    # 1d
+    pos1d = np.array([0, 0, 1, 1, 2])
+    out1d = compute_distances(pos1d)
+    assert isinstance(out1d, np.ndarray)
+    assert len(out1d) == pos1d.shape[-1] - 1
+    assert np.allclose(out1d, np.array([0, 1, 0, 1]))
 
-    out = compute_distances(xs, ys)
-    assert isinstance(out, np.ndarray)
-    assert len(out) == len(xs) -1
-    assert np.allclose(out, np.array([0, 1, 1, np.sqrt(2)]))
+    # 2d
+    pos2d = np.array([[0, 0, 1, 1, 2],
+                      [0, 0, 0, 1, 2]])
+    out2d = compute_distances(pos2d)
+    assert isinstance(out2d, np.ndarray)
+    assert len(out2d) == pos2d.shape[-1] - 1
+    assert np.allclose(out2d, np.array([0, 1, 1, np.sqrt(2)]))
 
 def test_compute_cumulative_distances():
 
-    xs = np.array([0, 0, 1, 1, 2])
-    ys = np.array([0, 0, 0, 1, 2])
+    # 1d
+    pos1d = np.array([0, 0, 1, 1, 2])
+    out1d = compute_cumulative_distances(pos1d)
+    assert isinstance(out1d, np.ndarray)
+    assert len(out1d) == pos1d.shape[-1] - 1
+    assert np.allclose(out1d, np.array([0, 1, 1, 2]))
 
-    out = compute_cumulative_distances(xs, ys)
-    assert isinstance(out, np.ndarray)
-    assert len(out) == len(xs) -1
-    assert np.allclose(out, np.array([0, 1, 1 + 1, 1 + 1 + np.sqrt(2)]))
+    # 2d
+    pos2d = np.array([[0, 0, 1, 1, 2],
+                      [0, 0, 0, 1, 2]])
[…]
