### 2023-05-25 50a0efc
add compute_distances_to_location func
