### 2023-05-25 923749f
add get_closest_position func
