+++ spiketools/tests/spatial/test_position.py
@@ -71,9 +71,9 @@ def test_compute_distances_to_location():
 
     # 1d
     pos1d = np.array([0, 0, 1, 1, 2])
-    point1d = 1
+    loc1d = 1
 
-    out1d = compute_distances_to_location(pos1d, point1d)
+    out1d = compute_distances_to_location(pos1d, loc1d)
     assert isinstance(out1d, np.ndarray)
     assert len(out1d) == pos1d.shape[-1]
     assert np.allclose(out1d, np.array([1, 1, 0, 0, 1]))
@@ -81,11 +81,29 @@ def test_compute_distances_to_location():
     # 2d
     pos2d = np.array([[0, 0, 1, 1, 2],
                       [0, 0, 0, 1, 2]])
-    point2d = [1, 0]
-    out2d = compute_distances_to_location(pos2d, point2d)
+    loc2d = [1, 0]
+    out2d = compute_distances_to_location(pos2d, loc2d)
     assert isinstance(out2d, np.ndarray)
     assert len(out2d) == pos2d.shape[-1]
     assert np.allclose(out2d, np.array([1, 1, 0, 1, np.sqrt(5)]))
+
+def test_get_closest_location():
+
+    # 1d
+    pos1d = np.array([0, 1, 2])
+
+    assert get_closest_position(pos1d, 1) == 1
+    assert get_closest_position(pos1d, 2) == 2
+    assert get_closest_position(pos1d, 1.75) == 2
+    assert get_closest_position(pos1d, 1.5, threshold=0.25) == -1
+
+    # 2d
+    pos2d = np.array([[0, 1, 2],
+                      [0, 0, 2]])
+    assert get_closest_position(pos2d, [1, 0]) == 1
+    assert get_closest_position(pos2d, [2, 2]) == 2
+    assert get_closest_position(pos2d, [1, 1]) == 1
+    assert get_closest_position(pos2d, [1, 1], threshold=0.25) == -1
 
 def test_compute_speed():
 
