### 2023-05-25 79798a0
reorg spatial distance & speed
