### 2023-05-27 25aa3e2
fix test function typo
