### 2023-05-27 fb36198
update docstrings & naming
