### 2023-05-30 79283b8
add dict helper functions
