### 2023-05-30 9a3672d
update listify to support arrays
