### 2023-05-30 4ea2906
update typechecking to deal with potential array inputs
