### 2023-06-02 #5268 @AJQuinn
Hi @TomDonoghue - can you confirm a final version and archive info for the JOSS tagged release of `spiketools`?

I think we can get through the post-review tasks quickly from there. Cheers
