### 2023-06-09 4f8a21a
add compute_trial_occupancy to api list
