### 2023-09-11 2a0d899
fix listify for multi-d arrays
