### 2023-09-11 aa6fe6b
update shuffle poisson to regenerate each shuffle
