### 2023-09-11 bf6d3e0
add more detailed notes to poisson shuffle func
