### 2023-09-11 e106976
move permute_vector to utils
