+++ spiketools/tests/stats/test_permutations.py
@@ -5,15 +5,6 @@ from spiketools.stats.permutations import *
 ###################################################################################################
 ###################################################################################################
 
-def test_permute_vector():
-
-    n_permutations = 5
-    data = np.array([1, 2, 3, 4, 5])
-
-    out = permute_vector(data, n_permutations=n_permutations)
-    assert isinstance(out, np.ndarray)
-    assert out.shape == (n_permutations, len(data))
-
 def test_compute_empirical_pvalue(tdata):
 
     p_value = compute_surrogate_pvalue(1.5, tdata)
