### 2023-09-11 a2e37f8
udpate tutorial for new dist funcs
