+++ spiketools/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.0.0'
+__version__ = '0.1.0rc1'
