### 2023-10-05 ec95067
switch thresholds to min / max values
