### 2023-10-05 7b955bd
update other funcs to reflect updates
