### 2023-10-07 24dc095
extend dim checking to take expected # values per sample
