### 2023-10-07 dba4105
use expected sample size across orientation inference
