### 2023-10-09 908460c
update make_orientation tests for columns
