+++ spiketools/utils/__init__.py
@@ -1,3 +1,3 @@
 """Utility functions."""
 
-from .utils import set_random_seed
+from .random import set_random_seed
