### 2023-10-09 93b1595
fix use of start_time in generator & sim func
