### 2023-10-09 212bc08
aesthetic updates to plot_rate_by_time
