+++ spiketools/plts/trials.py
@@ -68,7 +68,7 @@ def plot_rasters(data, vline=None, colors=None, vshade=None,
 
     ax.eventplot(data, colors=colors, **plt_kwargs)
 
-    add_vlines(vline, ax,
+    add_vlines(vline, ax, zorder=0,
                color=custom_plt_kwargs.pop('line_color', 'green'),
                lw=custom_plt_kwargs.pop('line_lw', 2.5),
                alpha=custom_plt_kwargs.pop('line_alpha', 0.5))
@@ -82,22 +82,24 @@ def plot_rasters(data, vline=None, colors=None, vshade=None,
 
 @savefig
 @set_plt_kwargs
-def plot_rate_by_time(x_vals, y_vals, average=None, shade=None, colors=None,
+def plot_rate_by_time(bin_times, trial_cfrs, average=None, shade=None, vline=None, colors=None,
                       labels=None, stats=None, sig_level=0.05, ax=None, **plt_kwargs):
     """Plot continuous firing rates across time.
 
     Parameters
     ----------
-    x_vals : 1d array
-        Values for the x-axis, for example time values or bin numbers.
-    y_vals : list of array
-        One or many set of values for the y-axis.
+    bin_times : 1d array
+        Values of the time bins, to be plotted on the x-axis.
+    trial_cfrs : list of array
+        Firing rate values, to be plotted on the y-axis.
         If each array is 1d values are plotted directly.
         If 2d, is to be averaged before plotting.
     average : {'mean', 'median'}, optional
         Averaging to apply to firing rate activity before plotting.
     shade : {'sem', 'std'} or list of array, optional
         Measure of variance to compute and/or plot as shading.
+    vline : float or list of float, optional
+        Location(s) to draw a vertical line. If None, no line is drawn.
     colors : str or list of str, optional
         Color(s) to plot the firing rates.
         If more than one, should match the number of conditions.
@@ -117,34 +119,42 @@ def plot_rate_by_time(x_vals, y_vals, average=None, shade=None, colors=None,
 
     ax = check_ax(ax, figsize=plt_kwargs.pop('figsize', None))
 
-    custom_kwargs = ['shade_alpha', 'legend_loc']
+    custom_kwargs = ['shade_alpha', 'legend_loc','line_color', 'line_lw', 'line_alpha']
     custom_plt_kwargs = get_kwargs(plt_kwargs, custom_kwargs)
 
     # If not a list of arrays, embed in a list to allow for looping (to support multiple inputs)
-    if not isinstance(y_vals[0], np.ndarray):
-        y_vals = [y_vals]
+    if not isinstance(trial_cfrs[0], np.ndarray):
+        trial_cfrs = [trial_cfrs]
     # If a 2d array that will get averaged to 1d, also embed into a list (not caught by above)
-    if isinstance(y_vals, np.ndarray) and y_vals.ndim == 2 and isinstance(average, str):
-        y_vals = [y_vals]
+    if isinstance(trial_cfrs, np.ndarray) and trial_cfrs.ndim == 2 and isinstance(average, str):
+        trial_cfrs = [trial_cfrs]
 
-    colors = DEFAULT_COLORS[0:len(y_vals)] if not colors else colors
+    colors = DEFAULT_COLORS[0:len(trial_cfrs)] if not colors else colors
     colors = [colors] if isinstance(colors, str) else colors
 
     if isinstance(shade, str):
-        shade = [get_var_func(shade)(arr, 0) for arr in y_vals]
+        shade = [get_var_func(shade)(arr, 0) for arr in trial_cfrs]
 
     if isinstance(average, str):
[…]
