+++ tutorials/plot_measures.py
@@ -11,6 +11,10 @@ This tutorial primarily covers the ``spiketools.measures`` module.
 # Applying measures & conversions to spiking data
 # -----------------------------------------------
 #
+# This tutorial explores functionality for converting between different representations and
+# applying measures to spike data. To show this functionality, we will use some example
+# simulated data.
+#
 # Sections
 # ~~~~~~~~
 #
@@ -43,8 +47,8 @@ from spiketools.sim import sim_spiketimes
 
 # Import plot functions
 from spiketools.plts.spikes import plot_isis
-from spiketools.plts.trials import plot_rasters
-from spiketools.plts.spatial import plot_heatmap
+from spiketools.plts.trials import plot_rasters, plot_rate_by_time
+#from spiketools.plts.spatial import plot_heatmap
 
 ###################################################################################################
 # Convert spiking data
@@ -86,9 +90,9 @@ print(spike_times[0:5])
 # Convert a vector of spike times in seconds to a binary spike train using sampling rate of 1000
 spike_train = convert_times_to_train(spike_times, fs=1000)
 
-# Print the first 20 spikes in the binary spike train
-print('Spike times:', spike_times[spike_times < (20 / 1000)])
-print('Corresponding spike train:', spike_train[:20])
+# Print the first 25 samples in the binary spike train
+print('Spike times:', spike_times[spike_times < (25 / 1000)])
+print('Corresponding spike train:', spike_train[:25])
 
 # Plot the first second of the spike times
 plot_rasters(spike_times[spike_times < 1], title='Raster of spike times')
@@ -111,13 +115,13 @@ plot_rasters(spike_times[spike_times < 1], title='Raster spike times from spike 
 
 ###################################################################################################
 #
-# Another representation of spike activity is the inter-spike intervals, which
+# Another representation of spike activity is the inter-spike interval, which
 # reflects the distribution of time-values between each spike.
 #
 # We can convert from the inter-spike intervals to spike times using the
 # :func:`~.convert_isis_to_times` function.
 #
-# Here we can see the converted data has the same raster plot as the original data, minus an offset.
+# Here we can see the converted data has the same raster plot as the original data.
 #
 
 ###################################################################################################
@@ -135,7 +139,7 @@ plot_rasters(spike_times[spike_times < 1], title='Raster spike times from ISIS')
 # Compute measures of spiking activity
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #
-# The firing rate of a neuron measures how fast the neuron is firing, providing
+# The firing rate of a neuron measures how fast the neuron is firing. The firing rate reflects
 # a measure of spiking activity in spikes / second, representing the rate of firing.
 #
 # We can use the :func:`~.compute_firing_rate` function to compute the firing rate
@@ -144,9 +148,9 @@ plot_rasters(spike_times[spike_times < 1], title='Raster spike times from ISIS')
 
 ###################################################################################################
 
-# Compute the spike rate from spike times
-spike_rate = compute_firing_rate(spike_times)
-print('The spike rate is {:1.3f} Hz'.format(spike_rate))
+# Compute the firing rate from spike times
+firing_rate = compute_firing_rate(spike_times)
+print('The firing rate is {:1.3f} Hz'.format(firing_rate))
 
 ###################################################################################################
 #
@@ -164,7 +168,7 @@ plot_isis(isis, bins=None, range=None, density=False, ax=None)
 
 ###################################################################################################
 #
-# Next, we can compute the coefficient of variation of ISIs that we just calculated.
+# Next, we can compute the coefficient of variation of the ISIs that we just calculated.
 #
 # The coefficient of variation measures the variability of the spiking activity.
 #
[…]
