+++ tutorials/plot_spatial.py
@@ -11,17 +11,21 @@ This tutorial primarily covers the ``spiketools.spatial`` module.
 # Apply spatial measures to spatial data
 # --------------------------------------
 #
+# This tutorial explores functionality for analyzing how spiking data relates to space,
+# including analysis such as place and target cells. To do so, we will use some example
+# simulated data.
+#
 # Sections
 # ~~~~~~~~
 #
 # This tutorial contains the following sections:
 #
-# 1. Compute and plot distances and speed using position
-# 2. Divide position in spatial bin edges
+# 1. Compute and plot distances and speed from position data
+# 2. Divide position in spatial bins
 # 3. Compute spatial bin assignment using spatial bin edges
-# 4. Compute time in each timestamp sample
+# 4. Compute time spent in each spatial bin
 # 5. Compute and plot occupancy and position counts
-# 6. Compute 2D and 1D spatial information
+# 6. Compute 1D and 2D spatial information
 #
 
 ###################################################################################################
@@ -68,13 +72,13 @@ timestamps = np.array([0, 1, 2, 3, 5, 6.5, 7.5, 8.5, 9.5, 11.5, 12.5, 13.5, 14.5
 # Compute and plot distances and speed using position
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #
-# Use x- and y-position to compute distance between two or more points,
-# cumulative distance traveled and speed.
+# First, we will explore some basic properties of the simulated position data, including
+# visualizing the data.
 #
 
 ###################################################################################################
 
-# Look at the range of our x and y positions
+# Calculate the range of the x and y position data
 ranges = compute_pos_ranges(position)
 
 print('The x-position ranges from {:1.1f} to {:1.1f}'.format(ranges[0][0], ranges[0][1]))
@@ -101,7 +105,10 @@ plot_position_by_time(timestamps, y_pos, alpha=1, ls='-', marker='x', color='tab
 
 ###################################################################################################
 #
-# With our position data, we can use some function to compute distance measures, including:
+# Next, we will use the x- and y-position values to compute distance between each set of points,
+# cumulative distance traveled, and speed.
+#
+# With our position data, the following functions can be used for this, including:
 #
 # - :func:`~.compute_distance`: which computes the distance between two points
 # - :func:`~.compute_distances`: which computes distances across a sequence of positions
@@ -126,23 +133,24 @@ speeds = compute_speed(position, timestamps)
 
 # Plot distance and speed measures
 ax1, ax2, ax3 = make_axes(3, 1, sharex=True, hspace=0.4, figsize=(8, 6))
+plt_kwargs = {'alpha' : 1, 'ls' : '-', 'marker' : 'x', 'markersize' : 10}
 
 # Plot distance traveled at each time
-plot_position_by_time(timestamps[1:], dist_traveled,
-                      ax=ax1, alpha=1, ls='-', marker='x', color='tab:pink', markersize=10,
-                      title='Distance traveled at each point',
-                      xlabel='time (t)', ylabel='speed (u/t)')
+plot_position_by_time(timestamps[1:], dist_traveled, color='tab:pink',
+                      title='Distance between each point',
+                      xlabel='Time (s)', ylabel='Distance',
+                      **plt_kwargs, ax=ax1)
 
 # Plot cumulative distance traveled per time
-plot_position_by_time(timestamps[1:], cumulative_dist_traveled,
-                      ax=ax2, alpha=1, ls='-', marker='x', color='tab:olive', markersize=10,
-                      title='Cumulative distance traveled at each point',
-                      xlabel='time (t)', ylabel='speed (u/t)')
+plot_position_by_time(timestamps, cumulative_dist_traveled, color='tab:olive',
[…]
