### 2023-10-10 a13c6b9
interim update of stats tut
