### 2023-10-10 bc95dcc
tweaks to measures tut
