### 2023-10-10 6ec1f75
update conversions to deal with different time_ranges / start_times
