### 2023-10-10 fc8c433
update tests for conversion updates
