### 2023-10-10 6dec1e1
udpate test objects for spike offset
