### 2023-10-10 d0b111d
update tests for new test obejcts
