### 2023-10-10 6db2349
add start_time to shuffle options, and update tests
