+++ spiketools/tests/stats/test_shuffle.py
@@ -10,14 +10,19 @@ from spiketools.stats.shuffle import *
 ###################################################################################################
 ###################################################################################################
 
-def test_shuffle_spikes(tspikes):
+def test_shuffle_spikes(tspikes, tspikes_offset):
 
+    n_shuffles = 2
     approaches = ['ISI', 'CIRCULAR', 'BINCIRC']
     kwargs = [{}, {'shuffle_min' : 10}, {}]
 
-    for approach, kwarg in zip(approaches, kwargs):
-        shuffled = shuffle_spikes(tspikes, approach=approach, n_shuffles=1, **kwarg)
-        assert isinstance(shuffled, np.ndarray)
+    for tdata in [tspikes, tspikes_offset]:
+        for approach, kwarg in zip(approaches, kwargs):
+            shuffled = shuffle_spikes(tdata, approach=approach, n_shuffles=n_shuffles, **kwarg)
+            assert isinstance(shuffled, np.ndarray)
+            assert len(shuffled) == n_shuffles
+            for el in shuffled:
+                assert len(el) == len(tdata)
 
 def test_drop_shuffle_range():
 
@@ -40,7 +45,7 @@ def test_drop_shuffle_range():
     toutput = np.array([0.5, 3.5, 3.9, 4.1, 5.4, 5.9])
     assert np.allclose(out, np.array([toutput + 1] * n_shuffles))
 
-def test_shuffle_isis(tspikes):
+def test_shuffle_isis(tspikes, tspikes_offset):
 
     shuffled = shuffle_isis(tspikes, n_shuffles=1)
     assert isinstance(shuffled, np.ndarray)
@@ -54,12 +59,17 @@ def test_shuffle_isis(tspikes):
     out2 = shuffle_isis(tspikes, n_shuffles=1)
     assert not np.array_equal(out1, out2)
 
+    # Test spikes with start time
+    shuffled_start = shuffle_isis(tspikes_offset, n_shuffles=1, start_time=-5)
+    assert tspikes.shape[-1] == shuffled_start.shape[-1]
+    assert np.min(shuffled_start) < 0
+
     # Test with more shuffles
     n_shuffles = 5
     out_many = shuffle_isis(tspikes, n_shuffles=n_shuffles)
     assert out_many.shape == (n_shuffles, len(tspikes))
 
-def test_shuffle_circular(tspikes):
+def test_shuffle_circular(tspikes, tspikes_offset):
 
     shuffled = shuffle_circular(tspikes, 10, n_shuffles=1)
     assert isinstance(shuffled, np.ndarray)
@@ -73,12 +83,17 @@ def test_shuffle_circular(tspikes):
     out2 = shuffle_circular(tspikes, 10, n_shuffles=1)
     assert not np.array_equal(out1, out2)
 
+    # Test start time
+    shuffled_start = shuffle_circular(tspikes_offset, 10, n_shuffles=1, start_time=-5)
+    assert tspikes.shape[-1] == shuffled_start.shape[-1]
+    assert np.min(shuffled_start) < 0
+
     # Test with more shuffles
     n_shuffles = 5
     out_many = shuffle_circular(tspikes, 10, n_shuffles=n_shuffles)
     assert out_many.shape == (n_shuffles, len(tspikes))
 
-def test_shuffle_bins(tspikes):
+def test_shuffle_bins(tspikes, tspikes_offset):
 
     shuffled = shuffle_bins(tspikes, n_shuffles=1)
     assert isinstance(shuffled, np.ndarray)
@@ -92,12 +107,17 @@ def test_shuffle_bins(tspikes):
     out2 = shuffle_bins(tspikes, n_shuffles=1)
[…]
