### 2023-10-10 437e2ed
fix doctest
