### 2023-10-11 b8a841e
add test that shuffle drop range works with neg start time
