### 2023-10-11 59e56b0
add trials sims to tut
