+++ _downloads/1ce28eb34b7d7dcc9b057f6d029f6e8c/plot_sims.py
@@ -11,14 +11,17 @@ This tutorial primarily covers the ``spiketools.sims`` module.
 # Simulating spiking data
 # -----------------------
 #
+# This tutorial explores functionality for simulating spiking activity, including simulating
+# spike times or spike trains, each with different available approaches.
+#
 # Sections
 # ~~~~~~~~
 #
-# This tutorial contains the following topics:
+# This tutorial covers the following topics:
 #
 # 1. Simulating spike times, based on specified distributions
 # 2. Simulating spike trains, based on spike probability or specified distributions
-# 3. Utilities for working with simulated spiking data
+# 3. Simulating trials of spiking data
 #
 
 ###################################################################################################
@@ -29,14 +32,15 @@ This tutorial primarily covers the ``spiketools.sims`` module.
 import numpy as np
 
 # Import simulation-related functions
-from spiketools.sim import sim_spiketimes, sim_spiketrain
+from spiketools.sim import sim_spiketimes, sim_spiketrain, sim_trials
 from spiketools.sim.times import sim_spiketimes_poisson
 from spiketools.sim.train import sim_spiketrain_binom, sim_spiketrain_poisson, sim_spiketrain_prob
+from spiketools.sim.trials import sim_trials_poisson
 from spiketools.sim.utils import apply_refractory_times, apply_refractory_train
 
 # Import plot functions
+from spiketools.plts.data import plot_lines
 from spiketools.plts.trials import plot_rasters
-from spiketools.plts.data import plot_hist
 
 # Import utilities
 from spiketools.measures.conversions import convert_train_to_times
@@ -62,7 +66,7 @@ spike_times_poisson = sim_spiketimes_poisson(rate=5, duration=2)
 plot_rasters(spike_times_poisson)
 
 ###################################################################################################
-# General function for simulation spike times
+# General function for simulating spike times
 # ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 #
 # In the above, we used a specific function for the spike time generation approach.
@@ -159,25 +163,29 @@ plot_rasters(spike_times_prob)
 #
 # In the above example, we specified a single spiking probability for all samples.
 #
-# We can also specify different spiking probability across samples,
+# We can also specify different spiking probabilities across samples,
 # such as in the following example.
 #
 
 ###################################################################################################
 
-# Simulate spike train of size 1000, based on a probability of spiking per sample
-p_spiking = np.random.random(1000)
-spike_train_prob_diff = sim_spiketrain_prob(p_spiking)
+# Simulate spike train of size 100, based on randomly sampled probability of spiking per sample
+n_probs = 100
+p_spiking = np.random.random(n_probs) / 10
+
+# Plot the continuous probabilities of spiking
+plot_lines(np.arange(n_probs), p_spiking)
 
 ###################################################################################################
+
+# Simulate spike train based on continuous probability of spiking
+spike_train_prob_diff = sim_spiketrain_prob(p_spiking)
 
 # Convert binary spike train to spike times in seconds
 spike_times_prob_diff = convert_train_to_times(spike_train_prob_diff, fs=1000)
 
-# Plot the empirical distribution of spike times probability across the simulated data
-plot_hist(spike_times_prob_diff, density=1, bins=50,
-          xlabel='Spike Times (seconds)', ylabel='Probability',
-          title='Probability Distribution of Spike Times')
+# Plot the spike times
+plot_rasters(spike_times_prob_diff)
 
 ###################################################################################################
 # Simulate spike trains based on probability distributions
@@ -192,28 +200,46 @@ plot_hist(spike_times_prob_diff, density=1, bins=50,
 #
 
 ###################################################################################################
+#
[…]
