+++ _downloads/3e1328885e3703795461db6650f78c35/plot_stats.py
@@ -11,7 +11,10 @@ This tutorial primarily covers the ``spiketools.stats`` module.
 # Applying statistical measures to spiking data
 # ---------------------------------------------
 #
-# This tutorial covers applying statistical measures to spike data.
+# This tutorial covers applying statistical measures to spiking data.
+#
+# Sections
+# ~~~~~~~~
 #
 # This tutorial contains the following sections:
 #
@@ -40,6 +43,7 @@ from spiketools.stats.trials import (compute_pre_post_ttest, compare_pre_post_ac
 # Import other spiketools functions
 from spiketools.sim.times import sim_spiketimes
 from spiketools.sim.train import sim_spiketrain_binom
+from spiketools.sim.trials import sim_trials_poisson
 from spiketools.measures.trials import compute_pre_post_rates
 from spiketools.spatial.occupancy import (compute_bin_edges, compute_bin_assignment,
                                           compute_bin_counts_assgn)
@@ -51,11 +55,11 @@ from spiketools.plts.utils import make_axes
 
 ###################################################################################################
 #
-# First we will generate some test data, by simulating some spike times.
+# First we will generate some test data, by simulating spike times.
 #
-# To do so, we will simulate some spike times at 10Hz for 100 seconds.
+# To do so, we will simulate spike times at 10Hz for 100 seconds.
 # In doing so, we will use the `refractory` settings to prevent multiple
-# spikes occurring within the same millisecond
+# spikes occurring too close together in time.
 #
 
 ###################################################################################################
@@ -67,7 +71,7 @@ spikes = sim_spiketimes(10, 100, 'poisson', refractory=0.001)
 # Compute and plot different shuffles of spikes
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #
-# Here, we will explore different shuffle approaches to create different shuffled
+# Here, we will explore different shuffling approaches to create different shuffled
 # surrogates of our original spikes.
 #
 # The approaches we will try are:
@@ -79,7 +83,7 @@ spikes = sim_spiketimes(10, 100, 'poisson', refractory=0.001)
 
 ###################################################################################################
 
-# Shuffle spikes using the four described methods
+# Shuffle spikes using the different methods - ISI, circular, and bins
 shuffled_isis = shuffle_isis(spikes, n_shuffles=10)
 shuffled_circular = shuffle_circular(spikes, shuffle_min=200, n_shuffles=10)
 shuffled_bins = shuffle_bins(spikes, bin_width_range=[0.5, 7], n_shuffles=10)
@@ -91,20 +95,11 @@ plot_rasters(spikes, xlim=[0, 6], title='Non-shuffled', vline=None)
 
 ###################################################################################################
 
-# Plot different shuffles
+# Plot the different shuffles - ISI, circular, and bin
 ax1, ax2, ax3 = make_axes(3, 1, sharey=True, hspace=0.3, figsize=(14, 9))
-
-# Shuffle spikes based on inter-spike intervals
-plot_rasters(shuffled_isis, xlim=[0, 6], ax=ax1,
-             title='Shuffle ISIS n_shuffles = 10')
-
-# Circular shuffle
-plot_rasters(shuffled_circular, xlim=[0, 6], ax=ax2,
-             title='Shuffle circular n_shuffles = 10')
-
-# Shuffled spikes using a binned circular shuffle
-plot_rasters(shuffled_bins, xlim=[0, 6], ax=ax3,
-             title='Shuffle bins n_shuffles = 10')
[…]

[…]
