### 2023-10-11 76973db
tutorial tweaks & fixes
