### 2023-10-11 5a6eba8
paper checks & tweaks
