+++ paper/paper.md
@@ -28,27 +28,27 @@ affiliations:
     index: 2
   - name: Department of Neurological Surgery, Columbia University
     index: 3
-date: 25 February, 2023
+date: 11 October, 2023
 bibliography: paper.bib
 ---
 
 # Summary
 
-A common method of collecting and analyzing neural activity is to implant electrodes that record the electrical activity of the brain, from which action potentials of individual neurons can be recorded [@buzsaki_origin_2012]. After pre-processing to detect spike waveforms and cluster them into groups representing putative single neurons [@rey_past_2015], this data can be used to investigate how neurons in the brain encode and process information. Analyzing single-unit activity requires dedicated analysis approaches, including representing spiking activity as spike times and/or binary spike trains, and analysis tools that allow for associating this activity to features of interest, for example the position of the animal in space or the properties of presented visual stimuli. To assist in this process, ``spiketools`` is a package designed to be used by neuroscientists for analyzing spiking activity.
+A common method of collecting and analyzing neural activity is to implant electrodes that record the electrical activity of the brain, from which action potentials of individual neurons can be recorded [@buzsaki_origin_2012]. After pre-processing to detect spike waveforms and cluster them into groups representing putative single neurons [@rey_past_2015], this data can be used to investigate how neurons in the brain encode and process information. Analyzing single-unit activity requires dedicated analysis approaches, including representing spiking activity as spike times and/or binary spike trains, and analysis tools that allow for associating this activity to features of interest, for example the position of the subject in space or the properties of presented visual stimuli. To assist in this process, ``spiketools`` is a package designed to be used by neuroscientists for analyzing spiking activity.
 
 ``spiketools`` is written in the Python programming language, built on existing tools in the scientific Python ecosystem, and developed using best-practice procedures. The module is accompanied by a [documentation site](https://spiketools.github.io/) that includes detailed [tutorials](https://spiketools.github.io/spiketools/auto_tutorials/index.html) for each of the modules, which are described below, as well as suggested workflows for combining them.
 
 Modules in ``spiketools`` include:
 
 * measures : measures and conversions that can be applied to spiking data
-* objects : objects that can be used to manage spiking data
+* objects : objects that can be used for managing spiking data
 * spatial : space related functionality and measures
 * stats : statistical measures for analyzing spiking data
 * sim : simulations of spiking activity and related functionality
-* plts : plotting functions for visualizing spike data and related measures
+* plts : plotting functions for visualizing spiking data and related measures
 * utils : additional utilities for working with spiking data
 
-``spiketools`` is has the following required dependencies:
+``spiketools`` has the following required dependencies:
 
 * numpy : used for managing and computing with arrays [@harris_array_2020]
 * scipy : used for some existing algorithms [@virtanen_scipy_2020]
@@ -57,17 +57,23 @@ Modules in ``spiketools`` include:
 
 ``spiketools`` also has some optional dependencies that offer extra functionality:
 
-* statsmodels : used for statistically testing responses [@seabold_statsmodels_2010]
+* statsmodels : used for additional statistical measures [@seabold_statsmodels_2010]
 
 # Statement of Need
 
-``spiketools`` is an open-source Python package for analyzing spiking neural data. Spiking neural activity is an idiosyncratic data stream with specific properties that requires specialized analysis tools and dedicated algorithms and statistical tools. Despite the popularity of this kind of data, there is currently a lack of openly available and maintained tools for this kind of data, especially within the Python ecosystem. ``spiketools`` therefore fills a niche, leveraging the power of the scientific power ecosystem, while providing dedicated implementations for the specific requirements of spiking data.
[…]
