+++ paper/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 # Summary
 
-A common method of collecting and analyzing neural activity is to implant electrodes that record the electrical activity of the brain, from which action potentials of individual neurons can be recorded [@buzsaki_origin_2012]. After pre-processing to detect spike waveforms and cluster them into groups representing putative single neurons [@rey_past_2015], this data can be used to investigate how neurons in the brain encode and process information. Analyzing single-unit activity requires dedicated analysis approaches, including representing spiking activity as spike times and/or binary spike trains, and analysis tools that allow for associating this activity to features of interest, for example the position of the subject in space or the properties of presented visual stimuli. To assist in this process, ``spiketools`` is a package designed to be used by neuroscientists for analyzing spiking activity.
+A common method of collecting and analyzing neural activity is to implant electrodes that record the electrical activity of the brain, from which action potentials of individual neurons can be recorded [@buzsaki_ephys_2012]. After pre-processing to detect spike waveforms and cluster them into groups representing putative single neurons [@rey_spikesorting_2015], this data can be used to investigate how neurons in the brain encode and process information. Analyzing single-unit activity requires dedicated analysis approaches, including representing spiking activity as spike times and/or binary spike trains, and analysis tools that allow for associating this activity to features of interest, for example the position of the subject in space or the properties of presented visual stimuli. To assist in this process, ``spiketools`` is a package designed to be used by neuroscientists for analyzing spiking activity.
 
 ``spiketools`` is written in the Python programming language, built on existing tools in the scientific Python ecosystem, and developed using best-practice procedures. The module is accompanied by a [documentation site](https://spiketools.github.io/) that includes detailed [tutorials](https://spiketools.github.io/spiketools/auto_tutorials/index.html) for each of the modules, which are described below, as well as suggested workflows for combining them.
 
@@ -50,7 +50,7 @@ Modules in ``spiketools`` include:
 
 ``spiketools`` has the following required dependencies:
 
-* numpy : used for managing and computing with arrays [@harris_array_2020]
+* numpy : used for managing and computing with arrays [@harris_numpy_2020]
 * scipy : used for some existing algorithms [@virtanen_scipy_2020]
 * pandas : used for managing heterogeneous data [@mckinney_pandas_2011]
 * matplotlib : used for plotting [@hunter_matplotlib_2007]
@@ -63,7 +63,7 @@ Modules in ``spiketools`` include:
 
 ``spiketools`` is an open-source Python package for analyzing spiking neural data. Spiking neural activity is an idiosyncratic data stream with specific properties that requires specialized analysis tools including dedicated algorithms and statistical tools. Despite the popularity of this kind of data, there is currently a lack of openly available and maintained tools for this kind of data, especially within the Python ecosystem. ``spiketools`` therefore fills a niche, leveraging the power of the scientific Python ecosystem, while providing dedicated implementations for the specific requirements of spiking data.
 
-Benefits of ``spiketools`` include that it follows modular organization, includes a test suite, follows a release cycle with versioned updates, and includes documentation and tutorials. ``spiketools`` is designed with a lightweight architecture in which functions take in arrays of spike times or spike trains, thus offering a flexible toolbox for custom analyses of spiking data. This approach also makes the tool flexible such that it can be integrated into existing codebases and workflows that use other tools.
[…]
