### 2023-10-12 02bbb52
add start_time as a common arg to shuffle
