+++ _downloads/3e1328885e3703795461db6650f78c35/plot_stats.py
@@ -37,8 +37,7 @@ import numpy as np
 from spiketools.stats.shuffle import shuffle_spikes, shuffle_isis, shuffle_circular, shuffle_bins
 from spiketools.stats.permutations import compute_surrogate_stats
 from spiketools.stats.anova import create_dataframe, create_dataframe_bins, fit_anova
-from spiketools.stats.trials import (compute_pre_post_ttest, compare_pre_post_activity,
-                                     compare_trial_frs)
+from spiketools.stats.trials import compute_pre_post_ttest, compare_pre_post_activity
 
 # Import other spiketools functions
 from spiketools.sim.times import sim_spiketimes
@@ -200,33 +199,6 @@ print('Average FR post-event: {:1.3f}'.format(avg_post))
 print('The t-value is {:1.3f}, with a p-value of {:.3e}'.format(tval_t, pval_t))
 
 ###################################################################################################
-# Compare 1st and 2nd half of trials
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#
-# In real data, time or trial index may influence the outcome.
-#
-# For another example analysis, next we will examine the binned pre / post firing
-# rates, split up by a split-half of the trials, to examine whether there is
-# evidence for a difference in spiking activity across different trials.
-#
-# To do this, we can use the :func:`~.compare_trial_frs` function.
-#
-
-###################################################################################################
-
-# Combine trials together into a 2d array
-all_trials_data = np.array([frs_pre, frs_post]).T
-
-# For each bin (pre and post event bins), compare 1st half of trials to 2nd half of trials
-pre_between_halfs, post_between_halfs = compare_trial_frs(all_trials_data[:5], all_trials_data[5:])
-
-# Print out the results
-print('For pre bin comparison, the t-value is {:1.3f}, a p-value of {:.3e}.'.format(\
-      pre_between_halfs[0], pre_between_halfs[1]))
-print('For post bin comparison, the t-value is {:1.3f}, with a p-value of {:.3e}.'.format(\
-      post_between_halfs[0], post_between_halfs[1]))
-
-###################################################################################################
 # Run 2-way ANOVA on multiple trials data
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #
@@ -320,8 +292,9 @@ print('Difference in post-pre firing rate of real data: {:1.2f}'.format(fr_diff)
 #
 # To compare the delta firing rate, we need to generate our distribution of surrogates.
 #
-# Thus, we will shuffle the data from the first trial (as an example) 100 times (using isis).
-# For each of the shuffles, we will calculate the change in firing rate.
+# To do so, we will shuffle the data such that each individual trial gets shuffled, to create
+# surrogates with randomly re-organized trial data. We can then compute our measure(s) of
+# interest on this surrogate data, to compare to the real data results.
 #
 
 ###################################################################################################
@@ -329,13 +302,14 @@ print('Difference in post-pre firing rate of real data: {:1.2f}'.format(fr_diff)
 # Settings for shuffling the data
 n_shuffles = 100
 
+# Create surrogate data and compute measures of interest
 shuffled_diffs = []
 for shuffle_ind in range(n_shuffles):
 
     # Shuffle the spike data - shuffle each trial to make a new set of shuffled trials
     shuffled_trials = []
     for t_ind in range(n_trials):
-        shuffled_trials.append(shuffle_isis(trial_spikes[t_ind], n_shuffles=1))
+        shuffled_trials.append(shuffle_isis(trial_spikes[t_ind], n_shuffles=1, start_time=-3))
 
[…]
