### 2023-10-12 #5268 @TomDonoghue
I don't seem to have the edit rights to click on the author tasks above, but I think I've done them all!

We have put through a couple final updates to spiketools, so that it's ready for a new release. I also did a quick check through of the paper, and made some minor updates.

There is a new tagged release of spiketools (v0.1.0), that is tagged through Github (https://github.com/spiketools/spiketools/releases), has an associated PYPI release (https://pypi.org/project/spiketools/), and is also tracked on Zenodo (https://zenodo.org/record/8436780).

I also confirm that I:
- Checked the authors, affiliations, and ORCIDs on the paper
- Release the new version 0.1.0 (as above)
- Archived this release on Zenodo: https://zenodo.org/record/8436780
- Made sure that the title and list of Authors on Zenodo matches the paper
- Made sure the Zenodo archive license listed matches that of the code

@AJQuinn and/or @arfon - is there anything else you need from us to finalize the JOSS paper? Please let us know!

Apologies again for the delays in getting this one completed!
