### 2023-10-13 #5268 @AJQuinn
Fantastic - thanks @TomDonoghue . I'll check over the remaining points

I'm not able to follow your zenodo link at the moment unfortunately - getting some odd errors, there is an announcement that the whole website is undergoing maintenance/bug-fixes after an update. I'll keep an eye on it and let you know if there is anything further to do.
