### 2023-10-16 #5268 @AJQuinn
Hi @TomDonoghue - Zenodo seems to be running more smoothly since the weekend, though I'm still getting a page not found error for the release, could you take a look? Perhaps their sever dropped your changes during the maintenance.
