### 2023-10-16 #5268 @TomDonoghue
Hey @AJQuinn - I sent in a service request to Zenodo a day or two ago and got a general reply that they're looking into the issue and hope to have things back to normal during the day today - so hopefully that addresses it soon! If it's not back up and normal by tomorrow morning, I'll try and do a new release.
