### 2023-10-19 ecdcc2b
udpate tests to test >0 start
