+++ spiketools/tests/stats/test_shuffle.py
@@ -10,13 +10,13 @@ from spiketools.stats.shuffle import *
 ###################################################################################################
 ###################################################################################################
 
-def test_shuffle_spikes(tspikes, tspikes_offset):
+def test_shuffle_spikes(tspikes, tspikes_offset_neg, tspikes_offset_pos):
 
     n_shuffles = 2
     approaches = ['ISI', 'CIRCULAR', 'BINCIRC']
     kwargs = [{}, {'shuffle_min' : 10}, {}]
 
-    for tdata in [tspikes, tspikes_offset]:
+    for tdata in [tspikes, tspikes_offset_neg, tspikes_offset_pos]:
         for approach, kwarg in zip(approaches, kwargs):
             shuffled = shuffle_spikes(tdata, approach=approach, n_shuffles=n_shuffles, **kwarg)
             assert isinstance(shuffled, np.ndarray)
@@ -32,33 +32,44 @@ def test_drop_shuffle_range():
         return np.array([spikes + 1] * n_shuffles)
 
     n_shuffles = 2
-    spikes = np.array([0.5, 1.5, 1.9, 4.1, 5.4, 5.9])
+    tspikes = np.array([0.5, 1.5, 1.9, 4.1, 5.4, 5.9])
 
     # Check without activating decorator
-    out = _shuffle_spikes(spikes, n_shuffles)
-    assert np.allclose(out, np.array([spikes + 1] * n_shuffles))
+    out = _shuffle_spikes(tspikes, n_shuffles)
+    assert np.allclose(out, np.array([tspikes + 1] * n_shuffles))
 
     # Check with applying a drop range
     #   This test the spikes that get moved from the empty range (spikes @ ind: [1, 2])
     drop_time_range = [2, 4]
-    out = _shuffle_spikes(spikes, n_shuffles, drop_time_range=drop_time_range)
+    out = _shuffle_spikes(tspikes, n_shuffles, drop_time_range=drop_time_range)
     toutput = np.array([1.5, 4.5, 4.9, 5.1, 6.4, 6.9])
     assert np.allclose(out, np.array([toutput] * n_shuffles))
 
-    # Test that this all still works with negative values / different start time
-    spikes_off = np.array([-1.5, -1.1, -0.5, 0.1, 0.7, 4.1, 5.1])
+    ## Test that this all still works with different start time: negative values
+    spikes_off_neg = np.array([-1.5, -1.1, -0.5, 0.1, 0.7, 4.1, 5.1])
 
     # Check without activating decorator
-    out_off_no_drop = _shuffle_spikes(spikes_off, n_shuffles)
-    assert np.allclose(out_off_no_drop, np.array([spikes_off + 1] * n_shuffles))
+    out_off_no_drop_neg = _shuffle_spikes(spikes_off_neg, n_shuffles)
+    assert np.allclose(out_off_no_drop_neg, np.array([spikes_off_neg + 1] * n_shuffles))
 
     # Check with activating decorator
-    drop_time_range_off = [1, 4]
-    out_off = _shuffle_spikes(spikes_off, n_shuffles, drop_time_range=drop_time_range_off)
-    toutput_off = np.array([-0.5, -0.1, 0.5, 4.1, 4.7, 5.1, 6.1])
-    assert np.allclose(out_off, np.array([toutput_off] * n_shuffles))
+    out_off_neg = _shuffle_spikes(spikes_off_neg, n_shuffles, drop_time_range=[1, 4])
+    toutput_off_neg = np.array([-0.5, -0.1, 0.5, 4.1, 4.7, 5.1, 6.1])
[…]
