### 2023-10-20 d2828da
update threshold_spikes_by_values to use min/max thresh
