### 2023-10-21 8aa2e78
update place calc to use min / max speed
