### 2023-10-21 4e5417c
fix doctest for thresh extract
