### 2023-10-21 11fda69
add tests for empty extractions
