### 2023-10-21 11e75b8
udpate get_values_by_times & tests for special case 1 val situation
