### 2023-10-21 8d93730
tweak compute_bin_edges for output type with no pos inpit
