### 2023-11-01 #5268 @AJQuinn
Hi @TomDonoghue - looks good to me.

Final tweak: can you double check the citation for pandas, there is no DOI or URL in the current cited version and it differs from the recommended citations on the pandas docs [https://pandas.pydata.org/about/citing.html](https://pandas.pydata.org/about/citing.html)

Is there a particular reason to use the 2011 proceeding without a link? Could you have a look at including or switching to the citations on the pandas docs with valid DOIs?
