+++ paper/paper.bib
@@ -8,7 +8,7 @@
     journal = {Nature Reviews Neuroscience},
     author = {Buzsáki, György and Anastassiou, Costas A. and Koch, Christof},
     year = {2012},
-    pages = {407--420}
+    pages = {407--420},
 }
 
 @article{buccino_spikeinterface_2020,
@@ -36,7 +36,6 @@
     year = {2019},
     pages = {1272},
 }
-
 
 @article{donoghue_singleneurons_2023,
     title = {Single neurons in the human medial temporal lobe flexibly shift representations across spatial and memory tasks},
@@ -85,6 +84,15 @@
     author = {McKinney, Wes},
     year = {2011},
     pages = {9},
+}
+
+@InProceedings{mckinney_pandas_2010,
+    title = {{Data} {Structures} for {Statistical} {Computing} in {Python}},
+    author = {McKinney, Wes},
+    booktitle = {{Proceedings} of the 9th {Python} in {Science} {Conference}},
+    ages = {56--61},
+    year = {2010},
+    doi = {10.25080/Majora-92bf1922-00a},
 }
 
 @article{garcia_neo_2014,
