### 2023-11-06 #5268 @TomDonoghue
Hey @AJQuinn - I updated the pandas citation (there was no good reason it was the one it was - I think I just found that one on google scholar without checking the recommended citation from the pandas docs, may bad).

Also, I've made a new release of spiketools (0.2.0), as we noticed a couple bugs in the 0.1.0 version. It has now been released on Github / Pypi, and has a new zenodo release: https://zenodo.org/records/10076024
If it's easy enough, the JOSS paper could be updated to refer to this version, though no worries if that's annoying - installing the module will get the new version anyways.
