### 2023-11-08 #5268 @AJQuinn
Hi @TomDonoghue looks good to me, happy to recommend to accept. A final check will be carried out by an Editor in Chief who may raise additional points if anything has been missed.

One point that might come up (not certain how strict the policy is here) is that the Zenodo repo name for v0.2.0 now doesn't match the paper title - you can update this in the metadata and might want to do that now.
