### 2023-11-08 #5268 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
