### 2023-11-09 ca71682
fix plt kwargs for plot_points
