### 2023-11-09 #5268 @Kevin-Mattheus-Moerman
@TomDonoghue as AEiC I am here to help process this work for acceptance in JOSS. I have check the paper, this review, the repository, and the archive link. Most seems in order. However, I do have the points below that require your attention:

- [ ] Please update the archive title to match the paper title: `spiketools: a Python package for analyzing single-unit neural activity`
- [ ] Please add the city and country to your affiliations e.g.  `New York, United States of America` (ensure you spell out the country, so do not use USA).
