### 2023-11-17 2b53ed7
Update README.rst
Add citation info
