+++ CITATION.cff
@@ -22,3 +22,28 @@ authors:
 repository-code: 'https://github.com/spiketools/spiketools'
 url: 'https://spiketools.github.io/'
 license: Apache-2.0
+preferred-citation:
+  type: article
+  title: 'spiketools: A Python package for analyzing single-unit neural activity'
+  authors:
+  - given-names: Thomas
+    family-names: Donoghue
+    orcid: 'https://orcid.org/0000-0001-5911-0472'
+  - given-names: 'Sandra'
+    family-names: 'Maesta Pereira'
+    orcid: 'https://orcid.org/0000-0001-6522-8311'
+  - given-names: 'Claire Z'
+    family-names: 'Han'
+    orcid: 'https://orcid.org/0000-0001-9710-8381'
+  - given-names: 'Salman E'
+    family-names: 'Qasim'
+    orcid: 'https://orcid.org/0000-0001-8739-5962'
+  - given-names: 'Joshua'
+    family-names: 'Jacobs'
+    orcid: 'https://orcid.org/0000-0003-1807-6882'
+  journal: 'Journal of Open Source Software'
+  doi: '10.21105/joss.05268'
+  issue: 91
+  volume: 8
+  year: 2023
+  start: 5268
