+++ doc/reference.rst
@@ -3,14 +3,29 @@ Reference
 
 This page describes how to reference and report on using this module.
 
-..
-.. Table of Contents
-.. -----------------
-.. .. contents::
-..    :local:
-..    :backlinks: none
+Table of Contents
+-----------------
+.. contents::
+   :local:
+   :backlinks: none
 
-.. Reference
-.. ~~~~~~~~~
+Referencing the Module
+~~~~~~~~~~~~~~~~~~~~~~
 
-If you use this module please cite the Github repository.
+If you use this code in your project, please cite:
+
+.. code-block:: text
+
+    Donoghue, T., Maesta-Pereira, S., Han C. Z., Qasim, S. E., & Jacobs, J. (2023)
+    spiketools: A Python package for analyzing single-unit neural activity.
+    Journal of Open Source Software, 8(91), 5268. DOI: 10.21105/joss.05268
+
+Direct Link: https://doi.org/10.21105/joss.05268
+
+Referencing Specific Methods
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+In some cases, a method or process available in spiketools may be a re-implementation
+that has a specific source and reference point, in which case the original should
+also be cited. In such cases, there will be a 'Reference' section in the function
+documentation, providing a reference to cite.
