### 2023-11-18 df343d0
minor doc updates
